refactor and fix some problems
